Deal or No Deal
In the game show, contestants pick a case out of 10 total cases, each of which
contains a different dollar amount printed inside. The dollar amount represents how much the
contestant can win at the end of the game if they make no ‘deals’ throughout the show. The
amounts in the cases are random, and the contestant has no idea which case has what amount.
Throughout the show, contestants pick out the other cases they did not originally pick one by
one, and the dollar amount inside those cases is revealed. Periodically, the ‘banker,’ a behind the
scenes person on the show, will come up with a dollar amount they will give the contestant on
the spot to make a ‘deal’ if the contestant chooses not to continue. This is largely based upon
which values have been revealed in the cases chosen. In other words, if several small cash
amounts have been revealed, it is more likely the contestants original chosen case contains a
large value, so the bank will make a larger cash offer for a deal. This continues until the
contestant makes a deal or decided to not make any deals and is ultimately given what is in their
original case.